AdminService/requirements.txt
sqlalchemy
quart
asyncmy
hypercorn
python-dotenv
mysqlclient
pymysql
cryptography
pydantic
aiokafka
argon2-cffi
argon2-cffi-bindings